+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Improved cargo docs
 - Created `CHANGELOG.md`
 - Adjusted visibility, default to `pub(crate)` when `pub` not needed
+- Add `call()` function to `Context`
 
 ## [0.1.1] – 2021-03-23